requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
pydantic==2.6.1 